+++ packages/graphql-codegen-cli/declarations.d.ts
@@ -7,3 +7,4 @@ declare module 'is-valid-path' {
 declare module 'listr-update-renderer';
 declare module 'indent-string';
 declare module 'json-to-pretty-yaml';
+declare module 'bdd-stdin';
