+++ packages/graphql-codegen-cli/package.json
@@ -83,6 +83,7 @@
     "@types/js-yaml": "3.12.0",
     "@types/listr": "0.13.0",
     "@types/log-symbols": "2.0.0",
+    "bdd-stdin": "0.2.0",
     "ts-jest": "24.0.0"
   },
   "main": "./dist/index.js",
