+++ packages/plugins/typescript-apollo-angular/tests/apollo-angular.spec.ts
@@ -4,7 +4,7 @@ import { plugin, addToSchema } from '../src/index';
 import { parse, GraphQLSchema, buildClientSchema, buildSchema, extendSchema } from 'graphql';
 import { DocumentFile } from 'graphql-codegen-core';
 import { plugin as tsPlugin } from '../../typescript/src/index';
-import { plugin as tsDocumentsPlugin } from '../../typescript-documents/src/index';
+import { plugin as tsDocumentsPlugin } from '../../typescript-operations/src/index';
 import { validateTs } from '../../typescript/tests/validate';
 import { readFileSync } from 'fs';
