+++ packages/plugins/typescript-operations/.gitignore
