+++ packages/plugins/typescript-operations/.npmignore
