+++ packages/plugins/typescript-operations/src/ts-selection-set-to-object.ts
